Added test_custom_variable_names function to ``test_core.py``
