+++ astropy/modeling/tests/test_core.py
@@ -1493,3 +1493,17 @@ def test_model_string_indexing():
 
     assert compound["Model1"] == gauss
     assert compound["Model2"] == airy
+
+
+def test_custom_variable_names():
+
+    """Test the possibility of setting custom variables names in models"""
+
+    g = models.Gaussian1D(amplitude=1.2, mean=0.5, stddev=0.3, name = "test_gaussian")
+
+    model_str = str(g)
+
+    expected_names = ["test_gaussian_amplitude", "test_gaussian_mean", "test_gaussian_stddev"]
+    for name in expected_names:
+        assert name in model_str, f"Custom variable name '{name}' not found in model string representation"
+
